WIP: DevTools: Remove text inputs from tab order when in dialog
Despite having tabindex=-1 (but no explicit tabindex set), DevTools <span>
inputs appear in the taborder. Normally this is desirable, but it trips up
the logic that removes hidden controls from the taborder when a dialog appears.
This CL explicitly sets and restores tabindex on these fields when they're
encountered by a dialog.

Change-Id: Ib392989b239ef70770769326b3ad89f9bb406925
Bug: 1271320