Store previously sent profiler messages for profiler panel
The profiler panel requires all events of CPUProfilerModel for profile
finishes. The previous solution was instantiating the profiler panel in
SDK and then would fire the listener. To remove a circular dependency
and unblock the removal of `profiler/module.json`, store these messages
instead and iterate through them when the profiler panel is
instantiated.

R=bmeurer@chromium.org

Bug: 1127902
Change-Id: Ia0d3056704776f2e30a0b843f421eda1f2d05957