Fix console.table
The legacy datagrid expects all users of its API to call `.asWidget`,
but doesn't actually enforce doing that. The `ConsoleViewMessage`
wasn't calling `.asWidget`, as it doesn't use a proper widget structure
itself.

Instead, we can manually recreate the widget with a shadowroot boundary,
so that the styles can properly be injected. Additionally, we need
to carry over the `objectValue` styles that `ConsoleView` also injects
in its root.

R=alexrudenko@chromium.org

Bug: 1271191
Change-Id: If9934ef54d51a961d8c0df7f02b96c7d9b0a5b92