[bfcache] Use chrome://terms instead of chrome://version
This CL replaces the URL to navigate to with the Run Test button with
chrome://terms, as chrome://version disables JavaScript and could lead
to a crash.

We considered about:blank, too, but navigation to about:blank is treated
as same site, and we have not shipped same-site bfcache to 100% yet, so
for now we are going to use chrome://terms.

Bug: 1269176
Change-Id: I64fd678b5d8fd626c89bcd70d1c2ed69c7ed33eb