Allow tests to remove throttler timeout for ServiceWorkerCacheModel
The test http/tests/devtools/cache-storage/cache-live-update-list.js
currently sets the _timeout of the throttler manually. This is error
prone as it reaches into the internal throttler state and no longer
works after the private field migration.

Instead, this CL adds a public method that changes the
Throttler in ServiceWorkerCacheModel to execute `schedule` calls
immediately. This method can be called by tests.

R=tvanderlippe@chromium.org

Bug: 1222126
Change-Id: I72e047bbab76bad0652a58fa47434a962b886bbb