Remove unit test folder exemption
Previously, we exempted our unit tests from importing only entrypoints.
However, we later realized that we always have to import the entrypoint
to avoid build non-determinism problems. Therefore, remove these exemptions,
which incidentally caught some existing violations.

The ColorUtils file is special in inspector_overlay, so we can safely
exempt that one as a one-off. The e2e-test need to import the helpers
as all other e2e-tests do. The hello-world is an example, so let's ignore
that one for now.

R=jacktfranklin@chromium.org

Bug: 1271490
Change-Id: Ib05d799b547e46cde4349b8738a5ecd63d50f3e1