[console] Fix fallback anchor display.
In the `Linkifier` we'd previously try to make the `anchor` look like
the `fallbackAnchor` in place when a target was disposed, but that
for example missed the event handlers (i.e. they were still in place
afterwards). Since we already have the `fallbackAnchor` constructed
it's straight forward to just put it in place of the `anchor` in the
DOM tree.

Also don't override the `cursor` for `.devtools-link` in console
messages, but stick to the (system wide) defaults, which ensures that
links that aren't supposed to be clickable (have the `preventClick`
option set to a trueish value) don't change to a `pointer` cursor
even inside the Console panel.

Fixed: chromium:1272413
Bug: chromium:1270248, chromium:1271078
Change-Id: Ib3ee739b1e4742667072a7c2d4ab4fd30847a1de