[cleanup] Consistently use ConsoleMessage.script when available.
The `ConsoleMessage` objects already carry a `scriptId` field these
days, but this wasn't set consistently, and was also not used
consistently. For example for `ConsoleMessage`s created with a
`stackTrace`, the `scriptId` wasn't always set, but instead in
arbitrary places we went fishing for script IDs on the call frames.

Drive-by-fix: Remove `ConsoleMessage.setExecutionContextId` and instead
just pass the `executionContextId` via the options bag to the
constructor (which already accepts that field anyways).

Bug: chromium:1271078
Change-Id: Ia47a8bde8ca593ca05c8932c6fc5aa3e33b1767c