Fix rendering layout shift details that don't have clusters
This code assumes that a layout shift in the Experience track will have
a `_current_cluster_id` and `_current_cluster_score`, however if we have
a trace with one LS, that `has_recent_input = true`, it never gets
assigned to a cluster (correctly). Therefore we need to render these two
bits of information conditionally, otherwise we error when we try to
render.

Bug: 1259606
Change-Id: Ib3af80e10616331d02c53dc701673a4d81b704ff