Fix a potential accessibility issue with Infobar
Currently, if an Infobar UI (such as the pretty print info bar)
in dt;fe has a detailed link ("Learn more"), clicking on the link
may lose the focus, which is an accessibility issue.

I had a hard time reproducing a pretty print info bar in Chromium.
That's why a screenshot is not provided here.

Bug: 1273129
Change-Id: I83fbde8aed1c9786670dfd93e75c969d701c6499