Fix keyboard event on ListWiget's cancel button
When the Enter keydown event is sent to the cancel button, there
are two event handlers to handle it. One is registered on keydown
event on this.element. The other is registered on click event on
this.cancelButton. Note that this.cancelButton is a child node
of this.element.

The first handler handles the event and consumes the event. It also
wrongly triggers committing the entered input. If there is no input,
committing silently fails. The click event to cancel is therefore
never triggered.

The solution is to define the keydown event on the editor-content
div node, which is a sibling of this.cancelButton.

Fixed: 1274086
Change-Id: Ica181e57058b2037d5c792f6e2a2805c84e4f4e9