Remove needless styles from LinearMemoryValueInterpreter
When doing some clean up of inspectorCommon usage, I noticed that this
class added it as an adopted stylesheet in the `constructor`. However,
the `connectedCallback` then ran and removed it as an adopted
stylesheet. Therefore we can remove this import.

I have loaded up the LMI in DevTools and confirmed it still looks OK.

Bug: none
Change-Id: I2a48c8ea21bed31da2bcf2ce39438998371eabdc