Do not add semicolon to the last declaration if there was none
Previously when we update style text, when there is no semicolon at the
end of the *last* CSS declaration in a declaration block, we would add
a semicolon. So the following modification happens:
```
div { color: red }
```
becomes
```
div { color: red; }
```
This would bring unnecessary changes to authored CSS styles, and the
change would also be registered in the Changes tab. This CL removes
this automatic appending (only to the *last* declaration, otherwise
it's invalid CSS) under the preciseChanges experiment.

Bug: chromium:1238818, chromium:1268754
Change-Id: Ib9567dc9c881bc5b42f5df0f984aaf99c5981958