Remove entrypoints/shell/shell.json
Shell no longer has any content, so we can remove it from all
entrypoints. We still want to share the contents of the meta files and
all, so keep these in. However, since we are no longer processing the
`shell.js` file `build_release_applications`, we can now fold in
`shell-meta-files.ts` into `shell.js`.

Ideally, we also change `shell` to be TypeScript-authored, but we can
only do that after all entrypoints that import it (in their respective
`.js` files) are TypeScript. So we will have to do that last.

R=jacktfranklin@chromium.org

Bug: 1127902
Change-Id: I47239286f6f1dbff63ce69cf8ca4c78bd4761030