Move main side-effect instantiation to entrypoints
To ease the migration off the `.json` files, we first need to move the
main instantiation to the correct timing. If we perform it too early,
runnables that we register later don't work (which we are going to do in
a follow-up CL). However, side-effect instantiation is obscure anyways,
so let's move them into the appropriate entrypoints so that we make this
explicit.

R=jacktfranklin@chromium.org

Bug: 1127902
Change-Id: I5686c446a371f0fc0bb4e784c705775973530f4b