Remove entrypoints/js_app/js_app.json
Since this was the last entrypoint that was still processed with
`build_release_applications`, GN now starts to error and claims the
action is unused (which it is). Therefore, we need to clean up this
action as well as all the GN variables it had been using. I will clean
up the actual Python implementation in a follow-up CL, to keep this one
as small as it can be.

R=jacktfranklin@chromium.org

Bug: 1127902
Change-Id: Ie4296885e6825d72a0b6c356971290337b680a71