Remove final few side-effects in platform/utilities.ts
The final few can be added to TestRunner.js which can expose them to the
global scope for the legacy layout tests. The other usages of the final
global have been removed from the legacy defs and ported in the code.

As such, we can now consider crbug.com/1050549 finalized and all globals
properly exported.

R=jacktfranklin@chromium.org

Fixed: 1050549
Change-Id: I9945ecb15e6324ee1982a81fbab669c63fb62512