Clear sinon stubs between each unit test
Consider the following unit tests:

```
describe.only('jack test', () => {
  const someObj = {
    foo() {
      return 2;
    },
  };
  it('does a thing', () => {
    const stub = sinon.stub(someObj, 'foo').callsFake(() => {
      return 5;
    })
    assert.strictEqual(someObj.foo(), 5); // this test PASSES
  });

  it('does another thing', () => {
    assert.strictEqual(someObj.foo(), 2); // this test FAILS
  });
});
```

The second test fails in our setup because the stub, although it was
created in the first unit test, survives between tests.

To fix this we can call `sinon.restore()`, which ensures that any
stubbed methods get cleaned between runs, and ensures that in the above
test cases both tests pass. `test_setup.ts` is run automatically between
each test, so by adding it to the `afterEach` we ensure each test starts
with a clean slate as far as sinon is concerned.

Fixed: 1275936
Change-Id: Ic218027979b7d5b323d297d90891d741f61d5d88