Add e2e-test for quoted key autocompletions in console
In https://crrev.com/c/3310606 I initially requested a unit test
to be added to verify the bug in crbug.com/1275553 doesn't regress
again. However, to reproduce the bug we need to evaluate expressions
and use these in the later autocompletions.

Therefore, let's add an e2e-test for this behavior and verify that both
single quoted and double quoted keys are working as expected.

R=jacktfranklin@chromium.org

Bug: 1275553
Change-Id: I822ac2a778cda8bf2f8d13e5f1f9671707fb7c9f