Non-fatal errors for DOM revealer
Certain user-causable scenarios can cause the DOM node revealer to not
work correctly. For example, `console.log($0)` a selected element, then
delete the element from the tree, and then right-click on the element in
the Console tool and click "Reveal in Elements". In this situation,
the node is detached from the DOM, and can't be revealed.

A couple of these use cases individually output error cases to the
console, but some don't do anything. This change uniformly transfers the
error via the `catch` handler of revealer, and has a single point of
output.

Change-Id: I1a1898279935485ae0a6f149fe6c3926707ee4b7