Allow links from deprecation issues
This allows all links going to `www.chromestatus/feature/xxx` to be
displayed in Issue descriptions. In addition several other links found
in deprecation issues[1] are whitelisted as well.

The test for SAB issues needed to be updated, because there are now 2
issues showing up. Previously the deprecation issue was not shown due
to causing an error when rendering it.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/frame/deprecation.cc?q=deprecation.cc

Fixed: 1275681
Change-Id: I9f175ff1d65b97f486005aebc89c94005f87376b