[TimelinePanel] Clear timer before removing the StatusPane
This stops the timer before we remove the status pane. Otherwise the
timer would try to update the status pane based on its parent, although
this element has already been removed from the DOM.

Also, rename 'hide' to 'remove' as this is what this method is
doing.

Fixed: chromium:1223418, chromium:1255679
Change-Id: Iac7496db7bba6ab352ad47925bea5e2bff9ac7a1