Remove platform/utilities.ts
The `assertNever` function was a TypeScript utility, so let's move it
into typescript-utilities.ts. The onload implementation is no longer
required. Since our HTML pages include the JS link as the last one and
the full page has been loaded, it is effectively a noop and always
immediately calls `callback`. Therefore, we can inline the callbacks and
remove some hoops.

R=jacktfranklin@chromium.org

Bug: none
Change-Id: Ib530cb12324820fab8371e18f102c39ef6fbce51