Fix player title UI issues
There have been a few issues here with player titles for a while,
including the icons not being the same size, strings being too long, and
sometimes not being as descriptive as they could be, even when better
data could be provided.

Here is a screenshot for reference. I can record a usage video too.
https://files.tedm.io/dtss/new_players_list.png

R=tvanderlippe

Change-Id: Ib063f31f0f60c3125180dc300c78c3fcbbe45038