[refactor] Decouple `sprintf()` and console formatting.
Previously `StringUtilities.sprintf()` was coupled together in a
somewhat weird way with the console formatting. This had several
downsides:

1. Refactoring to match the Console Standard is close to impossible with
   the existing mix and match, since it would also change `sprintf()` in
   surprising ways (i.e. recursive substitutions with %s).
2. `sprintf()` had some surprising super powers, for example, it would
   eat ANSI color escape sequences, and it would refuse to replace %%
   with %, and also don't apply the specifier precision consistently and
   correctly.

This CL removes the coupling and puts a dead simple and straightforward
`sprintf()` implementation in place that matches the Single UNIX Speci-
fication, and is thus sufficiently easy to reason about. This opens the
door to refactor the console formatting machinery to finally match the
Console Standard.

Drive-by-fix: Remove redundant `vsprintf()` and replace its only use
with a spread call to `sprintf()`.

Bug: chromium:1282076
Change-Id: I88ac86f541625fd30ecdd040bcb2b8bddc1b1758