[console] Align message formatting with Console Standard.
This introduces a dedicated `ConsoleFormat.format()` function, which
correctly implements the (potentially) recursive formatting of arguments
to console.log() and friends as specified in the Console Standard.

I intentionally did not extend the generic `StringUtilities.format()`
function, which was designed to be used in many different ways, and for
other users, it would be surprising to get the console behavior (which
is why we previously already disentagled the `StringUtilities.sprintf()`
function in http://crrev.com/c/3354896).

Bug: chromium: 1277944
Change-Id: If51cc9a2bbb7eccb088f3351045c05ef37628029
Fixed: chromium:1282076
Spec: https://console.spec.whatwg.org
Doc: https://bit.ly/v8-proper-console-type-conversions