Make auto dark setting enable `prefers-color-scheme: dark`
The auto dark mode is intended to always be used together with
`prefers-color-scheme: dark`. This change makes the setting in the
Renderer tab both enable auto dark and sets the prefered color scheme to
dark. To make it clear that these two settings are intertwined, we
disable the select box for prefers-color-scheme when auto dark mode is
enabled. This change also makes the auto dark setting to a checkbox to
simplify the UX.

https://imgur.com/a/79wNwc0

Change-Id: I72b99d3f3aabd6e05ebac2df5bd1e85b3c791312
Bug: none