Restructure scheduleUpdateInDebugger to allow for awaiting it
This CL restructures scheduleUpdateInDebugger to allow calling
functions to await the update. In follow up changes we will
need to properly await scheduleUpdateInDebugger to make sure
that an update (setting the breakpoint) has happened before
we can continue execution. Specifically, this is needed
such that we can make sure that we have set all the breakpoints
before we continue the execution on e.g. instrumentation
breakpoints.

Bug: chromium:1133307, chromium:1229541
Change-Id: I3ef2103ff03aaa02503a680b190c2d46148448de