[Quick Open] Fix the space display problem in TextPrompt
Origianlly it used a <span> under an <input>,  the <span> will have same content with <input> and the hint text is in <span>::after.
The display problem is only in mac, because in Mac, the <input> and  <span> texts are different size even their conputed font-size are same.

So the best solution is change the <span> to <input>, so both element will be exactly same.

Before: https://imgur.com/a/hSTHTW8
After: https://imgur.com/a/A40BSzD

Fixed: 1270754
Change-Id: I5ded34bd099accd93351801ba1de209bfbe51b26