Fix handling of asynchronously removed location
This fixes a bug where SourcePanel.executionLineChanged can revive stale
source code if `await liveLocation.uiLocation()` gets pre-empted by
deleting the source code.

This can happen on a reload of a page with a source mapped JavaScript
file while the execution is stopped in that JavaScript file. When
resetting the DebuggerModel, we first detach source maps. The detach
handler will schedule execution line update (among other things).  Then
the scripts are removed. Once the scheduled execution line update runs,
it will revive the source code with the execution line in
TabbedEditorContainer. When/if the script is actually loaded by the VM,
the front-end will then incorrectly skip loading the script contents
from the VM because its content is already present in
TabbedEditorContainer.  As a result, if the newly run script had
different content it won't be reflected in the source panel.

This patch solves the problem by skipping the execution line update
logic if the execution line location got discarded in the meantime.

Bug: chromium:508270
Change-Id: I338e5fda0b43edf8537d7eceea14940c9fa33238