Correctly indent DOM elements that wrap
Each DOM element is represented by a <li> element, which contains
a number of <span> and <div> elements that render the content. A
::before rule renders the expand triangle.

By putting the content <span> and <div> elements into a new <div>
container, we can use `display: flex` to correctly align
the content wrt expand triangle.

Fixed: 1016266
Change-Id: I956a5c9f15db9d38254c9abc11f54743f9ee9863