Enable no-floating-promises
All promises must now be awaited, or explicitly marked with `void`.
The void expression will discard the result of the promise, but is
used to make it explicit that the promise result can be ignored.

This should not have a performance impact on the ESLint presubmit,
as we already use the type information for TypeScript eslint. However,
we should monitor the times in the next few weeks and see if we run
into any issues.

R=jacktfranklin@chromium.org

Fixed: 1239617
Change-Id: I2b35806a832459dc1bc45fc75208e509e182f568