Remove patchColorText from FlameChart
This migrates the hardcoded color values in JS to the CSS equivalents,
and in so doing removes the need for patchColorText. In this instance it
also moves some of the colors out to the flameChart.css file rather than
using globally-available colors, since none of those are a close enough
match.

R=jacktfranklin@chromium.org

Bug: 1278738
Change-Id: I37fd4491fd62e1ae2b8a91045aa7ea64963763e6