Remove legacy patching code
Insofar as I can tell the code that applies the theme now obtains an
empty list from document.styleSheets (well, one which has a null href
and is therefore excluded). This means that the applyTheme function can,
for now at least, just apply the dark theme class if needed and return.

This, in turn, means that all the patching functions are no longer
required. This CL removes those functions. A future CL can remove some
of the indirection of applyTheme, perhaps, and make it possible for us
to restyle DevTools without the reload.

R=jacktfranklin@chromium.org

Bug: 1278738
Change-Id: I2a53306e8c30d4f98df79643d5b8bb4b42b0b1c7