Use GN build timestamps for copyright year
To avoid non-determinism on January 1st of a new year, we have
to use the build timestamp from GN to compute the current year.
The build timestamp will be the same for all bots, which ensures
that all files are built with the same year in the copyright header.

R=jacktfranklin@chromium.org

Fixed: 1283883
Change-Id: Ifbe581ddb65095ca6ef555625be11a370cb8f238