Introducing a DebuggerStateMachine
This is 1/3 CLs to make sure to synchronize the BreakpointManager
and the DebuggerModel. The DebuggerStateMachine will reflect
in what state the debuggerModel is at the moment, and which
actions can happen.

It has the following states:
Disabled,
StartingUp,   -- when Debugger.enable is called
Enabled,      -- when the response of enable returns
ReadyToPause, -- when all cached breakpoints are set in v8
ShuttingDown  -- when Debugger.disable is called
Disabled      -- when the response of disable returns

Bug: chromium:1229541, chromium:1133307
Change-Id: I630a16f6e99337674a1a0acf10b7f9928212cf92