Fix data grid column dragging
We use each column's html element width as source of truth, which is
truncated to integer pixels. When computing widths upon resizing,
we use floating point. Truncating on each step causes the columns
to continuously shift to the left.

To solve this, we use the computed floating point sizes as source
of truth instead.

R=jacktfranklin@chromium.org

Fixed: 1288182
Change-Id: If00ff6cc3ea2817c6b7c620c3ee8be1e40255d84