[debugger] Reverse storage order of scripts with the same url
Everytime V8 executes a script, we create a new "Script" object on
the DevTools side with a unique script id. The source URL of the
snippet is the same though, accross all invocations of the same
snippet. This can cause problems down the road when we try to
linkify source locations based solely on the sourceURL without a
script ID to help. This can be the case when trying to linkify a
raw Error#stack string.

The DebuggerModel keeps track of these scripts with a map that
stores all scripts with the same sourceURL. Historically, the
scripts in the list for a single sourceURL are stored in the
order in which the scripts were parsed. This means a user
of the debugger model implicitly prefers older over newer
scripts when trying to resolve a source URL.

This CL changes the storage order, so existing users of
the DebuggerModel will prefer newer scripts over older ones
when they share the same source URL. The result is that e.g.
repeated snippet executions that throw errors, will have more
accurate links.

Bug: 1277403
Change-Id: I25707efde6f4f9ab2826d563724354e2046aa8ef