Synchronizing the BreakpointManager and DebuggerModel
This CL is introducing a BreakpointSyncManager that
communicates between the BreakpointManager and the DebuggerModel
to ensure that all breakpoints are set before:

1. continuing execution on a newly attached target,
and 2. running a newly parsed script.

These synchronization points are necessary to make sure
that breakpoints that are already set, are actually hit.
The idea is that we pause the execution until we are
done with setting the breakpoints on the v8 side.

Bug: chromium:1229541, chromium:1133307
Change-Id: I2fcadb81abe9d20848a691c7cfe424f8e62bfa56