Port local overrides from CDP's `Network` domain to `Fetch` domain
The methods and events for request interception in the `Network` domain
are deprecated. As a byproduct of moving to the `Fetch` domain, for
intercepted and overridden resources DevTools now shows the actual
response headers received from the server instead of some minimal/dummy
values which were previously set in
InterceptedRequest.continueRequestWithContent().

Bug: 1288023
Change-Id: I94cd5dbe5250e9f8502683271098304241551574