Hide autocomplete dialog when resizing panels
With https://crrev.com/c/3370501 we are now using pointer events for
resizing DevTools panels. This had the inadvertent side-effect of
no longer closing the SuggestBox that is part of the sources
Watch panel.

I confirmed that reverting the listeners in the above CL to `mousedown`
would fix the issue, but it wasn't immediately clear why. I then
added a `debugger` statement to `SuggestBox.hide()` and inspected
the callstack. The callstack pointed to `GlassPane.onMouseDown`, which
is bound to `mousedown`.

By listening to both mouse and pointer events, we now correctly close
any panels that should be closed whenever the user clicks outside
them.

R=yangguo@chromium.org
CC=fluorescent.hallucinogen@gmail.com

Fixed: 1289579
Change-Id: I8baca036481a5b5f14bc8eee9b1121ec01828da2