Consistently apply a single mechanism for setting use_python3.
chromium_trybot was previously using python3 by passing in the
use_python3 value to recipe. When devtools/devtools-frontend was
switched to python3, it was done in a way that ignored the use_python3
value, causing chromium_trybot to switch back back to python2. This
restores chromium_trybot to python3 by using the same mechanism as
devtools/devtools-frontend.

Bug: 1256405
Change-Id: I20867867232e040487392687c6e3e065b2993dc6